Change the currency to 30,5

+++ src/Installs/migrations/2014_05_26_060000_create_module_fields_table.php
@@ -31,7 +31,7 @@ class CreateModuleFieldsTable extends Migration
             $table->boolean('unique')->default(false);
             $table->string('defaultvalue');
             $table->integer('minlength')->unsigned()->default(0);
-            $table->integer('maxlength')->unsigned()->default(0);
+            $table->bigInteger('maxlength')->unsigned()->default(0);
             $table->boolean('required')->default(false);
             $table->text('popup_vals');
             $table->integer('sort')->unsigned()->default(0);
